API Backend/requirements.txt
fastapi==0.115.6
pydantic==2.12.5
openai==1.51.2
python-dotenv==1.0.1
requests==2.32.3
passlib==1.7.4
python-jose==3.3.0
sqlalchemy==2.0.36
uvicorn==0.32.1
psycopg2-binary==2.9.11
asyncpg==0.31.0
python-multipart==0.0.17
authlib==1.3.2
httpx>=0.28.0
httpcore>=1.0.8
PyPDF2==3.0.1
pypdf==5.1.0
python-docx==1.1.2
itsdangerous==2.2.0
email-validator==2.2.0
gunicorn==23.0.0
bcrypt==3.2.2
greenlet==3.3.1
slowapi==0.1.9
supabase>=2.27.0
